Add spell-checking support via the Sonnet framework
Sonnet is a Tier 1 KDE Framework providing support for spell-checking.
We now include this framework as an optional dependency for the Qt5 version
of Quassel. Since Tier 1 frameworks have no dependencies other than Qt
itself, it should be easy to integrate on all platforms.

This commit adds spell-checking support for both the input and the topic
widget, and a settings page for configuring the spell-checking engine,
if Quassel is built against Qt5 and the Sonnet framework was found.

Note in the current version, the feature is broken, pending investigation!